fix datatype

+++ Tether.cs
@@ -110,7 +110,7 @@ namespace Palantir
         {
             try
             {
-                await (await Program.Client.GetChannelAsync(Convert.ToUInt16(PalantirEndpoint.ChannelID))).SendMessageAsync("The observed message couldn't be found. \nSet a channel using `>observe #channel`!");
+                await (await Program.Client.GetChannelAsync(Convert.ToUInt32(PalantirEndpoint.ChannelID))).SendMessageAsync("The observed message couldn't be found. \nSet a channel using `>observe #channel`!");
             }
             catch(Exception e)
             {
